feat: new algorithms service file
